eval does the simplest form of init, also moved it to eval.rs
